# tests/openapi.json
{
  "components": {
    "schemas": {
...
      "GlossaryFormat": {
        "enum": [
          "csv",
          "tsv",
          "xliff",
          "tbx",
          "tmx",
          "git_glossary",
          "sql_table",
          "json_pairs",
          "lanes_sqlite",
          "lanes_xml"
        ],
        "title": "GlossaryFormat",
        "type": "string"
      },
      "GlossaryReorderRequest": {
        "additionalProperties": false,
        "properties": {
          "ordered_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "minItems": 0,
            "title": "Ordered Ids",
            "type": "array"
          }
        },
        "required": [
          "ordered_ids"
        ],
        "title": "GlossaryReorderRequest",
        "type": "object"
      },
      "GlossaryToggleBody": {
        "additionalProperties": false,
        "description": "Optional toggle body: if enabled is not provided, the route flips state.",
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "title": "GlossaryToggleBody",
        "type": "object"
      },
      "GlossaryToggleRequest": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "GlossaryToggleRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "OmniScribe PDF OCR and document processing API",
    "title": "OmniScribe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/glossary/import/url": {
      "post": {
        "description": "Import a glossary fetched from a remote URL.\n\nAccepts BOTH the old query-parameter shape and the modern JSON\nbody (see module docstring). Format is taken from the explicit\nquery/body parameter first, then inferred from the URL's path\nsuffix. A fetch failure (DNS, timeout, non-2xx) is mapped to a\n502 with an ``ai_error`` code so the client can distinguish\n\"we could not reach the URL\" from a downstream parse failure.",
        "operationId": "import_glossary_from_url_api_glossary_import_url_post",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Url"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "in": "query",
            "name": "encoding",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Encoding"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/GlossaryFormat"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Glossary From Url",
        "tags": [
          "glossary"
        ]
      }
    },
...
    "/api/glossary/library/entries": {
      "get": {
        "description": "Browse / search entries across every library glossary.\n\nSupports a free-text ``q`` query and an optional ``source_id``\nfilter; ``limit`` and ``offset`` paginate. The merged view is\nused when ``source_id`` is omitted.",
        "operationId": "get_library_entries_api_glossary_library_entries_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Library Entries",
        "tags": [
          "glossary"
        ]
      }
    },
...
    "/api/glossary/library/reorder": {
      "post": {
        "description": "Persist a new display order for the glossary library.\n\nThe request body carries the full ordered id list; the service\nvalidates that every id is known and that no entry is missing\nbefore persisting the new ordering.",
        "operationId": "reorder_library_api_glossary_library_reorder_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlossaryReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reorder Library",
        "tags": [
          "glossary"
        ]
      }
    },
    "/api/glossary/library/{glossary_id}": {
      "delete": {
        "description": "Delete a glossary entry (and its entries) from the library.\n\nRemoves both the library metadata row and every term that was\ningested from that source; the operation is idempotent at the\nrow level (a missing id returns the standard not-found\nenvelope).",
        "operationId": "delete_library_entry_api_glossary_library__glossary_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "glossary_id",
            "required": true,
            "schema": {
              "title": "Glossary Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Library Entry",
        "tags": [
          "glossary"
        ]
      }
    },
    "/api/glossary/library/{glossary_id}/enable": {
      "post": {
        "description": "Set the enabled flag for a glossary entry.\n\nToggles whether the glossary is consulted during translation /\nOCR. The request body must carry the explicit ``enabled`` value\n(``true`` or ``false``) \u2014 there is no implicit-flip variant on\nthis endpoint.",
        "operationId": "toggle_library_entry_api_glossary_library__glossary_id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "glossary_id",
            "required": true,
            "schema": {
              "title": "Glossary Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlossaryToggleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Library Entry",
        "tags": [
          "glossary"
        ]
      }
    },
    "/api/glossary/library/{glossary_id}/entries": {
      "get": {
        "description": "Browse entries within a single glossary.\n\nSame shape as :func:`get_library_entries` but always scoped to\nthe path-parameter glossary; ``source_id`` is unnecessary here.",
        "operationId": "library_entries_api_glossary_library__glossary_id__entries_get",
        "parameters": [
          {
            "in": "path",
            "name": "glossary_id",
            "required": true,
            "schema": {
              "title": "Glossary Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Library Entries",
        "tags": [
          "glossary"
        ]
      }
    },
    "/api/glossary/library/{source_id}/toggle": {
      "post": {
        "description": "Flip-or-set the enabled flag for a glossary source.\n\nWhen the body is omitted the service flips the current value\n(handy for the Flutter \"tap to toggle\" UI); when a body is\nprovided the explicit ``enabled`` value wins.",
        "operationId": "toggle_source_api_glossary_library__source_id__toggle_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/GlossaryToggleBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Source",
        "tags": [
          "glossary"
        ]
      }
    },
...
    "/api/glossary/sources/{source_id}": {
      "delete": {
        "description": "Delete a glossary source by id.\n\nConvenience alias for the library-endpoint delete; the older\n``/api/glossary/sources/{id}`` route is preserved for clients\nthat have not migrated to ``/api/glossary/library/{id}``.",
        "operationId": "delete_source_api_glossary_sources__source_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Source",
        "tags": [
          "glossary"
        ]
      }
    },
...
  }
}